# Conflicts:
#	build.gradle
#	src/main/java/com/realive/RealiveApplication.java
#	src/main/java/com/realive/controller/customer/CustomerQnaController.java
#	src/main/java/com/realive/controller/product/ProductController.java
#	src/main/java/com/realive/controller/public_api/KakaoController.java
#	src/main/java/com/realive/controller/public_api/ProductViewController.java
#	src/main/java/com/realive/controller/seller/SellerController.java
#	src/main/java/com/realive/domain/review/ReviewReport.java
#	src/main/java/com/realive/repository/review/ReviewReportRepository.java
#	src/main/java/com/realive/repository/review/crud/ReviewCRUDRepository.java
#	src/main/java/com/realive/repository/seller/SellerRepository.java
#	src/main/java/com/realive/security/customer/CustomUserDetailsService.java
#	src/main/java/com/realive/security/customer/CustomerJwtAuthenticationFilter.java
#	src/main/java/com/realive/security/customer/JwtTokenProvider.java
#	src/main/java/com/realive/service/order/OrderServiceImpl.java
#	src/test/java/com/realive/RealiveApplicationTests.java
#	src/test/java/com/realive/admin/AdminSignupTest.java

+++ src/main/java/com/realive/domain/auction/Bid.java
@@ -23,13 +23,11 @@ public class Bid extends BaseTimeEntity {
     private Integer auctionId;
 
     @Column(name = "customer_id")
-    private Integer customerId;
+    private Long customerId;
 
     @Column(name = "bid_price")
     private Integer bidPrice;
 
     @Column(name = "bid_time")
     private LocalDateTime bidTime;
-
-
 }
